Move isSvg_newLine test to common
