+++ coil-svg/src/androidInstrumentedTest/kotlin/coil3/decode/SvgDecoderTest.android.kt
@@ -1,8 +1,7 @@
-package decode
+package coil3.decode
 
 import coil3.BitmapImage
 import coil3.ImageLoader
-import coil3.decode.*
 import coil3.fetch.SourceFetchResult
 import coil3.request.Options
 import coil3.size.Dimension
@@ -18,7 +17,6 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
 import kotlinx.coroutines.test.runTest
-import okio.Buffer
 import okio.BufferedSource
 import okio.FileSystem
 import okio.buffer
@@ -50,21 +48,6 @@ class SvgDecoderTest {
         assertNull(decoderFactory.create(result, Options(context), ImageLoader(context)))
         assertFalse(source.exhausted())
         assertEquals(8192, source.buffer.size) // should buffer exactly 1 segment
-    }
-
-    /** Regression test: https://github.com/coil-kt/coil/issues/1154 */
-    @Test
-    fun isSvg_newLine() {
-        val text = "<svg\n" +
-            "   xmlns:dc=\"http://purl.org/dc/elements/1.1/\"\n" +
-            "   xmlns:cc=\"http://creativecommons.org/ns#\"\n" +
-            "   xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\"\n" +
-            "   xmlns:svg=\"http://www.w3.org/2000/svg\"\n" +
-            "   xmlns=\"http://www.w3.org/2000/svg\"\n" +
-            "   xmlns:inkscape=\"http://www.inkscape.org/namespaces/inkscape\"/>"
-        val buffer = Buffer().apply { writeUtf8(text) }
-
-        assertTrue(DecodeUtils.isSvg(buffer))
     }
 
     @Test
